BKRS-362-ci-tests-improvments
