+++ .github/dependabot.yml
@@ -10,7 +10,7 @@ updates:
     target-branch: dev
     schedule:
       interval: "weekly"
-    open-pull-requests-limit: 0
+    open-pull-requests-limit: 10
 
   - package-ecosystem: github-actions
     directory: /
