+++ codecov.yml
@@ -1,3 +1,2 @@
 ignore:
   - "examples"
-  - "models"
