+++ go.mod
@@ -1,6 +1,8 @@
 module github.com/aerospike/backup-go
 
-go 1.25.8
+go 1.25.0
+
+toolchain go1.25.8
 
 require (
 	cloud.google.com/go/storage v1.64.0
