+++ io/storage/azure/blob/reader_writer_test.go
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//go:build integration
+
 package blob
 
 import (
