+++ tools.go
@@ -1,0 +1,24 @@
+// Copyright 2024 Aerospike, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build tools
+
+// Package tools pins the versions of the build-time tools this repository uses,
+// so that "go install" and CI resolve the same version as "go.mod".
+// It is never part of a normal build.
+package tools
+
+import (
+	_ "github.com/vektra/mockery/v3"
+)
